i'm sorry for my self again

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,9 +11,9 @@
  * Return:the number of nodes
 */
 
-size_t print_list(const list_t *h)
+int print_list(const list_t *h)
 {
-size_t count = 0;
+int count = 0;
 
 
 while (h != NULL)
